Add SSL verify support
Added --ssl-verify cli option for the IPAM driver, which allows a user
to specify whether to enforce SSL certificate verification for the
connection to Infoblox. The default is false.

Alternatively, the name of a cerficate file in PEM format can be specified,
in whcih case, the certificate will be used to verify the Infoblox
certificate.

Minor refactoring to better handle SSL certificate verification error.