change project structure as per go package structure
Each go binary should be in its own package. If it is not the
case then we end up having multiple main functions in one package
and then go get fails with error "...main redeclared in this block
previous declaration...". This patch makes the two binaries,
ipam-driver and create-ea-defs in their own packages.

Also start with the unit tests for the ipam_driver.go