Switch to V2 plugin architecture (#33)
* change project structure as per go package structure

Each go binary should be in its own package. If it is not the
case then we end up having multiple main functions in one package
and then go get fails with error "...main redeclared in this block
previous declaration...". This patch makes the two binaries,
ipam-driver and create-ea-defs in their own packages.

Also start with the unit tests for the ipam_driver.go

* change project structure as per go package structure

Each go binary should be in its own package. If it is not the
case then we end up having multiple main functions in one package
and then go get fails with error "...main redeclared in this block
previous declaration...". This patch makes the two binaries,
ipam-driver and create-ea-defs in their own packages.

Also start with the unit tests for the ipam_driver.go

* Add support for v2 plugin

* V2 IPAM plugin

* Fix README.md and config.json

* Fix README.md formatting

* Fix Readme formatting

* Fix Readme formatting

* Add support for environment variables

* Use caarlos0/env to parse env variables

Also added log messages and updates to README.md

* Remove unused launch.sh

* Add CLI arguments documentation in README.md

Add the usage of configuration option passed via
command line to the plugin

* Fix README with proper configuration options